move tests for unauthed endpoints to own file
